+++ R/StoxAnalyticalBaselineFunctions.R
@@ -340,7 +340,7 @@ extractIndividualDesignParametersStoxBiotic <- function(StoxBioticData, Stratifi
 #' @concept StoX-functions
 #' @concept Analytical estimation
 #' @md
-DefineIndividualSamplingParameters <- function(processData, StoxBioticData, DefinitionMethod=c("SRS", "Stratified", "LengthStratified"), Parameters=c(), LengthInterval=numeric(), StratificationColumns=character(), UseProcessData=FALSE){
+DefineIndividualSamplingParameters <- function(processData, StoxBioticData, DefinitionMethod=c("SRS", "Stratified", "LengthStratified"), Parameters=character(), LengthInterval=numeric(), StratificationColumns=character(), UseProcessData=FALSE){
 
   #May want to expose this option if DefinitionMethods are added that only provides relative selection probabilities.
   CollapseStrata=FALSE
@@ -447,10 +447,10 @@ DefineSamplingHierarchy <- function(StoxBioticData, IndividualSamplingParameters
 #'  \code{\link[RstoxFDA]{PSUSamplingParametersData}}, and checks that all responding PSUs are present in data records.
 #'  
 #'  After correcting for non-response, the SamplingUnitId in \code{\link[RstoxFDA]{PSUSamplingParametersData}} will be replaced
-#'  by an ID (argument 'DataRecordId') so that sampling units can be brought into correspondance with how they are identified in lower
+#'  by an ID (argument 'DataRecordId') so that sampling units can be brought into correspondence with how they are identified in lower
 #'  level sampling (\code{\link[RstoxFDA]{IndividualSamplingParametersData}})
 #'  
-#'  If any respondants (rows of the SelectionTable of PSUSamplingParametersData that does not have NA for SamplingUnitId) are not
+#'  If any respondents (rows of the SelectionTable of PSUSamplingParametersData that does not have NA for SamplingUnitId) are not
 #'  found in 'SamplingUnitId', execution halts with an error.
 #'  
 #'  Response after selection can generally be considered a process that modifies the sampling parameters that are set by design.
@@ -475,7 +475,7 @@ DefineSamplingHierarchy <- function(StoxBioticData, IndividualSamplingParameters
 #' @concept Analytical estimation
 #' @md
 #' @export
-AssignPSUSamplingParameters <- function(PSUSamplingParametersData, StoxBioticData, SamplingUnitId, DataRecordId, DefinitionMethod=c("MissingAtRandom")){
+AssignPSUSamplingParameters <- function(PSUSamplingParametersData, StoxBioticData, SamplingUnitId=character(), DataRecordId=character(), DefinitionMethod=c("MissingAtRandom")){
   checkMandatory(PSUSamplingParametersData, "PSUSamplingParametersData")
   checkMandatory(StoxBioticData, "StoxBioticData")
   checkMandatory(DataRecordId, "DataRecordId")